+++ barelymusician/gain_processor.cpp
@@ -8,16 +8,16 @@ namespace barelyapi {
 namespace {
 
 // Gain threshold of -96db in amplitude.
-constexpr float kGainThreshold = 2e-5f;
+constexpr double kGainThreshold = 2e-5;
 
 // Unity gain in amplitude.
-constexpr float kUnityGain = 1.0f;
+constexpr double kUnityGain = 1.0;
 
 // Total ramp duration in seconds.
-constexpr float kUnityRampDurationSeconds = 0.05f;
+constexpr double kUnityRampDurationSeconds = 0.05;
 
 // Applies constant `gain`.
-void ApplyConstantGain(float gain, float* buffer, int num_channels,
+void ApplyConstantGain(double gain, float* buffer, int num_channels,
                        int num_frames) noexcept {
   if (std::abs(gain - kUnityGain) < kGainThreshold) {
     return;
@@ -27,20 +27,20 @@ void ApplyConstantGain(float gain, float* buffer, int num_channels,
     return;
   }
   for (int i = 0; i < num_channels * num_frames; ++i) {
-    buffer[i] *= gain;
+    buffer[i] *= static_cast<float>(gain);
   }
 }
 
 // Applies linear ramp of `num_ramp_frames` from `gain` to `target_gain`.
-float ApplyLinearRamp(float gain, float target_gain, int num_ramp_frames,
-                      float* buffer, int num_channels,
-                      int num_frames) noexcept {
-  const float ramp_increment_ =
-      (target_gain - gain) / static_cast<float>(num_ramp_frames);
+double ApplyLinearRamp(double gain, double target_gain, int num_ramp_frames,
+                       float* buffer, int num_channels,
+                       int num_frames) noexcept {
+  const double ramp_increment_ =
+      (target_gain - gain) / static_cast<double>(num_ramp_frames);
   for (int frame = 0; frame < std::min(num_ramp_frames, num_frames); ++frame) {
     gain += ramp_increment_;
     for (int channel = 0; channel < num_channels; ++channel) {
-      buffer[num_channels * frame + channel] *= gain;
+      buffer[num_channels * frame + channel] *= static_cast<float>(gain);
     }
   }
   return (num_ramp_frames <= num_frames) ? target_gain : gain;
@@ -49,7 +49,7 @@ float ApplyLinearRamp(float gain, float target_gain, int num_ramp_frames,
 }  // namespace
 
 GainProcessor::GainProcessor(int sample_rate) noexcept
-    : num_unity_ramp_frames_(static_cast<float>(sample_rate) *
+    : num_unity_ramp_frames_(static_cast<double>(sample_rate) *
                              kUnityRampDurationSeconds) {}
 
 void GainProcessor::Process(float* buffer, int num_channels,
@@ -76,6 +76,6 @@ void GainProcessor::Process(float* buffer, int num_channels,
   }
 }
 
-void GainProcessor::SetGain(float gain) noexcept { target_gain_ = gain; }
+void GainProcessor::SetGain(double gain) noexcept { target_gain_ = gain; }
 
 }  // namespace barelyapi
